chore: Add expense form popup
Add a popup form for adding expenses to the home page. The form includes fields for amount, category, date, and user paid. The form is hidden by default and can be opened by clicking the "+" button. The form can be closed by clicking the "x" button. This change improves the user experience by providing a convenient way to add expenses.